updated styling on checkout component
